content-security-policy/to-port-from-blink/inline-style-allowed-while-cloning-objects-expected.txt

CONSOLE ERROR: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-SKwGvORdKBYTYiM4lxIkanDyKH8J0qJ5Ix8LGkKsbhw='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 77: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-E8Yv3p3nivzZKSVWxu7y2Pr-w7Nco7lvx8r5nBVoLFA='), or a nonce ('nonce-...') is required to enable inline execution.

PASS successfullyParsed is true

TEST COMPLETE
This test ensures that styles can be set by object.cloneNode()
PASS node1.style.background is "yellow"
PASS node2.style.background is "yellow"
PASS node3.style.background is "blue"
PASS node4.style.background is "blue"
PASS node1.style.color is "red"
PASS node2.style.color is "red"
PASS node3.style.color is "green"
PASS node4.style.color is "green"
PASS window.getComputedStyle(node1).background is window.getComputedStyle(node2).background
PASS window.getComputedStyle(node3).background is window.getComputedStyle(node4).background
PASS window.getComputedStyle(node1).color is window.getComputedStyle(node2).color
PASS window.getComputedStyle(node3).color is window.getComputedStyle(node4).color
PASS ops.style.background is ""
getComputedStyle(clonedOps).background: rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box
PASS ops.style.color is "red"
PASS clonedOps.style.background is ""
PASS violetOps.style.background is "rgb(238, 130, 238)"
PASS window.getComputedStyle(clonedOps).background is window.getComputedStyle(ops).background
PASS window.getComputedStyle(clonedOps).color is window.getComputedStyle(ops).color
getComputedStyle(violetOps).background: rgb(238, 130, 238) none repeat scroll 0% 0% / auto padding-box border-box
PASS window.getComputedStyle(ops).background is not window.getComputedStyle(violetOps).background
PASS window.getComputedStyle(clonedOps).background is not window.getComputedStyle(violetOps).background
PASS ops.id is "ops"
PASS ops.id is clonedOps.id
This is a div (nodes)
This is a div. (node 1 or 2)
This is a div. (node 1 or 2)
This is a div. (node 3 or 4)
Node #4
Yet another div.
Yet another div.
Yet another div.
